Synthesizer: Apply AudioOutput settings to PatchPlay action (so it also plays stereo).

+++ Business/JJ.Business.Synthesizer/Extensions/DeleteRelatedEntitiesExtensions.cs
@@ -54,8 +54,11 @@ namespace JJ.Business.Synthesizer.Extensions
                 repositories.PatchRepository.Delete(patch);
             }
 
-            document.AudioOutput.UnlinkRelatedEntities();
-            repositories.AudioOutputRepository.Delete(document.AudioOutput);
+            if (document.AudioOutput != null)
+            {
+                document.AudioOutput.UnlinkRelatedEntities();
+                repositories.AudioOutputRepository.Delete(document.AudioOutput);
+            }
 
             foreach (AudioFileOutput audioFileOutput in document.AudioFileOutputs.ToArray())
             {
